% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hive.R
\docType{methods}
\name{HiveUber}
\alias{HiveUber}
\alias{dbConnect,HiveUberDriver-method}
\title{Hive Uber DBI wrapper}
\usage{
HiveUber()

\S4method{dbConnect}{HiveUberDriver}(drv, url, ...)
}
\arguments{
\item{drv}{An object created by \code{HiveUber()}}

\item{url}{the jdbc connection url}

\item{...}{Other options}
}
\description{
Hive Uber DBI wrapper

Authentication credentials are can be specified inside the url, or kerberos can be used.
}
\details{
The \code{hiveuber.subjectcreds} option sets a JVM flag necessary for kerberos 
authentication. See \code{zzz.R} for details.
}
\examples{
\dontrun{
require(DBI)
con <- dbConnect(HiveUber(), url="jdbc:hive2://host:port/schema")
dbListTables(con)
dbGetQuery(con, "Select count(*) from nfultz.iris")
}
}
\seealso{
\href{https://cwiki.apache.org/confluence/display/Hive/HiveServer2+Clients#HiveServer2Clients-ConnectionURLs}{Hive Wiki} for more connections options.
}
