% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hive.R
\docType{class}
\name{HiveUberDriver-class}
\alias{HiveUberDriver-class}
\alias{HiveUberDriver}
\title{HiveUber driver class.}
\usage{
\S4method{initialize}{HiveUberDriver}(.Object, ...)
}
\description{
HiveUber driver class.

Constructor of HiveUberDriver
}
\keyword{internal}
